+++ src/py42/sdk/queries/fileevents/file_event_query.py
@@ -1,5 +1,6 @@
 from py42._internal.compat import str
 from py42.sdk.queries import BaseQuery
+from py42.settings import security_events_per_page
 from py42.sdk.queries.query_filter import create_filter_group
 from py42.sdk.queries.query_filter import create_query_filter
 from py42.sdk.queries.query_filter import QueryFilterStringField